+++ common/changes/@autorest/test-utils/tests-backlog_2021-02-03-20-50.json
@@ -1,11 +1,11 @@
 {
   "changes": [
     {
-      "packageName": "@autorest/testing",
+      "packageName": "@autorest/test-utils",
       "comment": "Initial content, include file snapshot matching",
       "type": "minor"
     }
   ],
-  "packageName": "@autorest/testing",
+  "packageName": "@autorest/test-utils",
   "email": "tiguerin@microsoft.com"
 }
