+++ packages/extensions/modelerfour/test/setupJest.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line unicorn/filename-case
-import "@autorest/testing/dist/matchers";
+import "@autorest/test-utils/dist/matchers";
