+++ packages/libs/oai2-to-oai3/package.json
@@ -51,7 +51,7 @@
     "@types/source-map": "0.5.0",
     "jest": "^26.6.3",
     "ts-jest": "^26.4.4",
-    "@autorest/testing": "~0.0.1"
+    "@autorest/test-utils": "~0.0.1"
   },
   "dependencies": {
     "@azure-tools/datastore": "~4.1.0",
