+++ packages/libs/oai2-to-oai3/test/setup-jest.ts
@@ -1,1 +1,1 @@
-import "@autorest/testing/dist/matchers";
+import "@autorest/test-utils/dist/matchers";
